# default_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "API documentation for my project",
    "version": "1.0.0",
    "title": "My API Documentation",
    "contact": {
      "name": "John Doe",
      "url": "http://example.com",
      "email": "john@example.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
...
    "/admin/log/page": {
      "get": {
        "tags": [
          "日志管理"
        ],
        "summary": "queryPage",
        "operationId": "queryPageUsingGET_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "isAsc",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "module",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "operationType",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "operatorName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageNo",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDTO«LogPageVO»",
              "originalRef": "PageDTO«LogPageVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/admin/notification/detail/{id}": {
      "get": {
        "tags": [
          "通知管理"
        ],
        "summary": "queryNotificationDetail",
        "operationId": "queryNotificationDetailUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/NotificationDetailVO",
              "originalRef": "NotificationDetailVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/notification/page": {
      "get": {
        "tags": [
          "通知管理"
        ],
        "summary": "queryPage",
        "operationId": "queryPageUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "isAsc",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "pageNo",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDTO«NotificationPageVO»",
              "originalRef": "PageDTO«NotificationPageVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/role/create": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "创建角色或系统人员",
        "operationId": "createRoleUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "roleCreateDTO",
            "description": "roleCreateDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RoleCreateDTO",
              "originalRef": "RoleCreateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R",
              "originalRef": "R"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/role/delete": {
      "delete": {
        "tags": [
          "角色管理"
        ],
        "summary": "删除系统人员",
        "operationId": "deleteAdminsUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R",
              "originalRef": "R"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/admin/role/detail/{id}": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "获取当前管理人员的信息",
        "operationId": "editRoleAndAdminDetailUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PermissionDetailVO",
              "originalRef": "PermissionDetailVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/role/edit/{id}": {
      "put": {
        "tags": [
          "角色管理"
        ],
        "summary": "编辑系统人员",
        "operationId": "editRoleUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "roleEditDTO",
            "description": "roleEditDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/修改角色",
              "originalRef": "修改角色"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R",
              "originalRef": "R"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/role/list": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "获取权限列表",
        "operationId": "getPermissionListUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/PermissionDTO",
                "originalRef": "PermissionDTO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/role/page": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "分页",
        "operationId": "queryPageUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "adminId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "isAsc",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "pageNo",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDTO«RolePageVO»",
              "originalRef": "PageDTO«RolePageVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/role/roleList": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "获取角色列表",
        "operationId": "roleListUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/RoleInfoVO",
                "originalRef": "RoleInfoVO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/admin/role/status/{id}": {
      "put": {
        "tags": [
          "角色管理"
        ],
        "summary": "状态",
        "operationId": "statusUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/R",
              "originalRef": "R"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/admin/user/detail/{userId}": {
      "get": {
        "tags": [
          "管理端用户管理"
        ],
        "summary": "管理端查询用户详情",
        "operationId": "queryUserDetailUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UserDetailVO",
              "originalRef": "UserDetailVO"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/reconciliation/page": {
      "get": {
        "tags": [
          "对账管理"
        ],
        "summary": "queryPage",
        "operationId": "queryPageUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "isAsc",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "meterNo",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "pageNo",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "minimum": 1,
            "exclusiveMinimum": false,
            "format": "int32"
          },
          {
            "name": "reconciliationNo",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "reconciliationStatus",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "date"
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "userType",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PageDTO«ReconciliationPageVO»",
              "originalRef": "PageDTO«ReconciliationPageVO»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "LogPageVO": {
      "type": "object",
      "properties": {
        "createTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "description": {
          "type": "string",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "ip": {
          "type": "string",
          "refType": null
        },
        "module": {
          "type": "string",
          "refType": null
        },
        "operationType": {
          "type": "string",
          "refType": null
        },
        "operatorName": {
          "type": "string",
          "refType": null
        },
        "status": {
          "type": "string",
          "refType": null
        }
      },
      "title": "LogPageVO"
    },
...
    "NotificationDetailVO": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string"
        },
        "createTime": {
          "type": "string",
          "format": "date-time"
        },
        "expireTime": {
          "type": "string",
          "format": "date-time"
        },
        "level": {
          "type": "string"
        },
        "senderName": {
          "type": "string"
        },
        "senderRole": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      },
      "title": "NotificationDetailVO"
    },
    "NotificationPageVO": {
      "type": "object",
      "properties": {
        "content": {
          "type": "string",
          "refType": null
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "expireTime": {
          "type": "string",
          "format": "date-time",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "level": {
          "type": "string",
          "refType": null
        },
        "readStatus": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "title": {
          "type": "string",
          "refType": null
        },
        "type": {
          "type": "string",
          "refType": null
        }
      },
      "title": "NotificationPageVO"
    },
...
    "PageDTO«LogPageVO»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LogPageVO",
            "originalRef": "LogPageVO"
          },
          "example": [
            {
              "createTime": "2023-06-10T08:25:13",
              "description": "更新用户信息",
              "id": 5012,
              "ip": "192.168.1.100",
              "module": "用户管理",
              "operationType": "UPDATE",
              "operatorName": "admin",
              "status": "成功"
            },
            {
              "createTime": "2023-06-10T08:20:45",
              "description": "创建新电表",
              "id": 5011,
              "ip": "192.168.1.100",
              "module": "电表管理",
              "operationType": "INSERT",
              "operatorName": "admin",
              "status": "成功"
            }
          ]
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "example": 126
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "example": 1256
        }
      },
      "title": "PageDTO«LogPageVO»"
    },
...
    "PageDTO«NotificationPageVO»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NotificationPageVO",
            "originalRef": "NotificationPageVO"
          },
          "example": [
            {
              "content": "系统将于6月15日00:00-02:00进行升级维护，期间服务不可用",
              "createTime": "2023-06-10T10:00:00",
              "expireTime": "2023-06-15T02:00:00",
              "id": 1023,
              "level": "重要",
              "readStatus": 1,
              "title": "系统维护通知",
              "type": "系统通知"
            },
            {
              "content": "6月电费账单已生成，请及时缴纳",
              "createTime": "2023-06-05T00:00:00",
              "expireTime": "2023-06-25T23:59:59",
              "id": 1022,
              "level": "一般",
              "readStatus": 1,
              "title": "电费账单通知",
              "type": "账单通知"
            }
          ]
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "example": 8
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "example": 78
        }
      },
      "title": "PageDTO«NotificationPageVO»"
    },
...
    "PageDTO«ReconciliationPageVO»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReconciliationPageVO",
            "originalRef": "ReconciliationPageVO"
          },
          "example": [
            {
              "balance": 450.75,
              "reconciliationNo": 10023,
              "status": "已批准",
              "userId": 10025,
              "userType": "居民用户",
              "username": "张小明"
            },
            {
              "balance": 825.30,
              "reconciliationNo": 10022,
              "status": "已批准",
              "userId": 10026,
              "userType": "商业用户",
              "username": "李明"
            }
          ]
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "example": 5
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "example": 42
        }
      },
      "title": "PageDTO«ReconciliationPageVO»"
    },
    "PageDTO«RolePageVO»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RolePageVO",
            "originalRef": "RolePageVO"
          },
          "example": [
            {
              "account": "admin",
              "adminId": 1001,
              "createTime": "2023-01-01T00:00:00",
              "role": "系统管理员",
              "roleDesc": "系统管理员，拥有所有权限",
              "status": 1
            },
            {
              "account": "operator",
              "adminId": 1002,
              "createTime": "2023-01-10T09:00:00",
              "role": "运营人员",
              "roleDesc": "负责日常运营工作",
              "status": 1
            }
          ]
        },
        "pages": {
          "type": "integer",
          "format": "int64",
          "example": 2
        },
        "total": {
          "type": "integer",
          "format": "int64",
          "example": 5
        }
      },
      "title": "PageDTO«RolePageVO»"
    },
...
    "PermissionDTO": {
      "type": "object",
      "properties": {
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PermissionDTO",
            "originalRef": "PermissionDTO"
          },
          "refType": "PermissionDTO",
          "example": [
            {
              "permissionId": 1002,
              "permissionName": "用户管理-查看"
            },
            {
              "permissionId": 1003,
              "permissionName": "用户管理-编辑"
            }
          ]
        },
        "permissionId": {
          "type": "integer",
          "format": "int64",
          "refType": null,
          "example": 1001
        },
        "permissionName": {
          "type": "string",
          "refType": null,
          "example": "用户管理"
        }
      },
      "title": "PermissionDTO"
    },
    "PermissionDetailVO": {
      "type": "object",
      "properties": {
        "account": {
          "type": "string",
          "example": "admin"
        },
        "permissionList": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PermissionDTO",
            "originalRef": "PermissionDTO"
          },
          "example": [
            {
              "permissionId": 1001,
              "permissionName": "用户管理",
              "children": [
                {
                  "permissionId": 1002,
                  "permissionName": "用户管理-查看"
                },
                {
                  "permissionId": 1003,
                  "permissionName": "用户管理-编辑"
                }
              ]
            },
            {
              "permissionId": 2001,
              "permissionName": "电表管理",
              "children": [
                {
                  "permissionId": 2002,
                  "permissionName": "电表管理-查看"
                },
                {
                  "permissionId": 2003,
                  "permissionName": "电表管理-编辑"
                }
              ]
            }
          ]
        },
        "roleDesc": {
          "type": "string",
          "example": "系统管理员，拥有所有权限"
        },
        "roleName": {
          "type": "string",
          "example": "系统管理员"
        }
      },
      "title": "PermissionDetailVO"
    },
    "R": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "业务状态码，200-成功，其它-失败"
        },
        "data": {
          "type": "object",
          "description": "响应数据"
        },
        "msg": {
          "type": "string",
          "example": "OK",
          "description": "响应消息"
        },
        "requestId": {
          "type": "string",
          "example": "1af123c11412e",
          "description": "请求id"
        }
      },
      "title": "R",
      "description": "通用响应结果"
    },
...
    "RoleCreateDTO": {
      "type": "object",
      "properties": {
        "account": {
          "type": "string",
          "refType": null
        },
        "isRole": {
          "type": "boolean",
          "refType": null
        },
        "password": {
          "type": "string",
          "refType": null
        },
        "permissionIdList": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "refType": "integer"
        },
        "role": {
          "type": "string",
          "refType": null
        },
        "roleDesc": {
          "type": "string",
          "refType": null
        }
      },
      "title": "RoleCreateDTO"
    },
    "RoleInfoVO": {
      "type": "object",
      "properties": {
        "icon": {
          "type": "string",
          "example": "admin-icon"
        },
        "label": {
          "type": "string",
          "example": "系统管理员"
        },
        "value": {
          "type": "string",
          "example": "ADMIN"
        }
      },
      "title": "RoleInfoVO"
    },
...
    "UserDetailVO": {
      "type": "object",
      "properties": {
        "accountStatus": {
          "type": "string",
          "example": "正常"
        },
        "address": {
          "type": "string",
          "example": "北京市海淀区中关村南路10号3单元501室"
        },
        "contractEndDate": {
          "type": "string",
          "format": "date-time",
          "example": "2024-04-30T23:59:59"
        },
        "contractStartDate": {
          "type": "string",
          "format": "date-time",
          "example": "2023-05-01T00:00:00"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "example": 10025
        },
        "idCardNo": {
          "type": "string",
          "example": "110101199001011234"
        },
        "lastMeterReadingDate": {
          "type": "string",
          "format": "date-time",
          "example": "2023-06-01T10:30:00"
        },
        "lastPaymentDate": {
          "type": "string",
          "format": "date-time",
          "example": "2023-06-05T14:32:45"
        },
        "meterId": {
          "type": "string",
          "example": "20230001"
        },
        "outstandingBill": {
          "type": "integer",
          "format": "int32",
          "example": 0
        },
        "phone": {
          "type": "string",
          "example": "13812345678"
        },
        "userType": {
          "type": "string",
          "example": "居民用户"
        },
        "username": {
          "type": "string",
          "example": "张小明"
        }
      },
      "title": "UserDetailVO"
    },
...
    "修改角色": {
      "type": "object",
      "properties": {
        "account": {
          "type": "string",
          "refType": null
        },
        "password": {
          "type": "string",
          "refType": null
        },
        "role": {
          "type": "string",
          "refType": null
        }
      },
      "title": "修改角色"
    }
  }
}